Batch submit
[US866186 1710] Add the state_audit_interval_ms to the
IntegrityMonitor.properties

[US866186 1710] Added comments to the state audit property
(state_audit_interval_ms) which clarify how to turn it off.

[ECOMPD2TD-000] Update version.properties and pom.xml for iter 41

[ECOMPD2TD-000] Update version.properties and pom.xml for iter 41

Revert "[ECOMPD2TD-000] Update version.properties and pom.xml for
iter 41"

This reverts commit 61367d4c87fa6b153f707361537d9f802fb95a6e.

[US250837 1710] Completed first cut of logic changes to
PMStandbyStateChangeNotifier so that each segment of logic is only
executed if a state change actually occurred.

[US250837 1710] First cut on the JUnit for PMStandbyStateChangeNotifier.
At this point I am simply checking the operation by inspecting the logs.

[US250837 1710] Fixed the JUnit so it is not failing and added a note
on how to run the JUnit.

[US250837 1710] Added a method to get previousStandbyStatus from
PMStandbyStateChangeHandler and began using it in JUnit tests.

[US250837 1710] Finished the StandbyStateManagementTest.  This uses the
previousStandbyStatus value to test the function.

[US250837 1710] Clean up of PMStandbyStateChangeNotifier and JUnit test
to make static string values that are used throughout.

[US250837 1710] Found a couple strings that I missed when converting to
static variables.

[US250837 1710] Updated PMStandbyStateChangeNotifier and
StandbyStateManagementTest

[US250837 1710] Added a dedicated db audit persistence unit
auditDroolsPU to taylor the classes which will be audited.

[ECOMPD2TD-1309] Clean up old persistence data

This is an addition to 'policy-persistence' that cleans up
'sessioninfo' records with a 'lastmodificationdate' that is older than
a specified time (default = 900 seconds = 15 minutes).

[ECOMPD2TD-000] Correction to pom.xml versions

Conflicts:

	version.properties

Change-Id: I810d358c07e05fd36fa435a6fbb9e538e78b5011
Signed-off-by: Ralph Straubs <rs8887@att.com>