Update patch set 2
Patch Set 2:

(1 comment)

It is a way to add new capabilities, or alternative implementations of functionality into the system without 1) code cluttering or 2) breaking the base.   They are basically services which uses the ServiceLoader framework, and it encourages modularization for developers in a non-modular system (just to say in contrast with osgi, which modularization is at its core).

One example is policy-healthcheck/, that provides clearly distinct functionality in its own module without cluttering other areas.  It also provides and alternative implementation to healthchecks that the one present in policy-persistence.   

They depend on well defined interception points at which points the PolicyEngine invokes its features.   The dispatching should happen at critical transition points on the PolicyEngine lifecycle.

These capabilities are optional and can be enabled/disabled without affecting the rest of the code via an "options" shell script in $POLICY_HOME/bin directory.

Patch-set: 2
Label: Code-Review=0