Update patch set 1
Patch Set 1:

Pam, agree with try-with-resources and in fact, I thought about it and decided against it just to go the quickest route.   Right now, my goal is to get rid of warnings in the very least intrusive way, requiring minimum testing, since I am not that familiar with this area.   In a later exercise, we can consider refactoring some portions of the code.   Definitely, policy-persistence module, requires serious consideration for refactoring.

Patch-set: 1