Update patch set 3
Patch Set 3:

Ralph, I had the same concern, I decided when I wrote it to provide higher programability tooling over tighter system aspects control as guiding principle at this stage (for now, this may change as we get more mature).   This is though, part of a more generic problem, in general, with any interception architecture.   For example, when performing a callback, the calling thread can be hijacked, or take arbitrarily long times, or exit the system, just to say ..   Then need to do the invokations from thread pools, measure feature invokation times, etc .. as countermeasure.   We can revisit in the future.   At least for the time being, we should take the assumption that loaded modules are trusted and safe for execution, they have been previously tested and validated.

Patch-set: 3