Update patch set 1
Patch Set 1:

I have no comments as we have discussed this internally before releasing the patch.   Something that I would like at some point, if time allows, is an enhancement as we have discussed to get the liveness status of the thread model, and perhaps the flavor of it (may be classname or something of that sort, that gives an indication if it has been overridden by a feature) so can be reported through telemetry (REST) APIs.   If you were to agree Ralph, I would create a (Enhacement) "Task" in JIRA to track it, and assign to you.  But first, let me know if I have your agreement in that there is value in this enhancement?

Patch-set: 1
Label: Code-Review=0