[POLICY-30] optional loadable eelf feature
for backwards compatibility purposes or if desired to
use EELF and common-framework (from policy-common) for
logging purposes in miscellaneous environments.   It supports
EELF/Common Framework style logging at the same time than
traditional logging.

This functionality is an optional loadable module.

policy-docker changes need to be made to optionally
unpack to make it active (in a different submission).

Change-Id: I5076c462409f03415ce54f0c9b5f50b78abb18f9
Signed-off-by: Jorge Hernandez <jh1730@att.com>