Update patch set 1
Patch Set 1:

The 'configure_component' function would still need to be in both places, because it is called elsewhere within 'docker-install.sh'. If you wanted to share this function as well, the functions could be placed in a shared library, where both 'features' and 'docker-install.sh' could reference them.

Patch-set: 1
Label: Code-Review=0