Update patch set 2
Patch Set 2:

right I know that's all in the base ..

yes, that could be the case, for the scenario that you're pointing out with POLICY_HOME ..  for the other ones such as JAVA_HOME, KEYSTORE.. seems pretty unlikely.

There is though the custom "install/enable" and "install/disable" scripts that the feature owner could write that will be executed.    May be same approach could be implemented here support a feature specific "install/install" and "install/uninstall" invokation, if it exists under the install directory, on an operation basis.   If that's the case, the feature owner could plug in, whatever they want provided they have the environment.

So, one possibility if custom processing besides a simpler version of "configure_component" is to delegate it to the feature owner to do it in their custom install ..  

what do you think?

Patch-set: 2
Label: Code-Review=0