Update patch set 1
Patch Set 1:

Thanks Ralph, I'll bring it up with internal ATT folks to make sure mysql client (mariadb) is covered as a prereq.   They will also need to do python as well to get the telemetry shell to get feature parity with onap side.

I'll also change the "warning" to "error" in the message as we talked, effectively the script blocks the enabling of the feature so it is desired behavior.   I'll go ahead and merge, you should see in my next review the echo statement changed then.

Patch-set: 1