Update patch set 2
Patch Set 2:

Pam, will take a look on the junits, I think it is bringing up and tear down http servers on the fly with the junits, may run into this type of situations where trying to bind immediately after.   This does not seem to happen in local environments, wondering if there is some degree of parallizations on running the junits.  Nevertheless has to be fixed.  I'll look into it, and should have nothing to do with this review, will be fixed independently.

Patch-set: 2
Label: Code-Review=0