Eliminate hanging when updating Drools containers
'PolicyContainer.updateToVersion(...)' now hangs unless the Drools session
threads are stopped prior to the update. This wasn't an issue in the past,
and it isn't clear what causes it to happen now. It also happened in the
Junit tests I am currently working on.

In any case, the 'updateToVersion' method has now been changed to stop
the threads prior to an update attempt, do the update, and then restart
the threads.

Change-Id: I577a67c6e99a68bcbe3c2f476f93c0a72b93451a
Issue-ID: POLICY-234
Signed-off-by: Ralph Straubs <rs8887@att.com>