Update patch set 1
Patch Set 1:

Drew, there is some overlap between the FileSystemPersistenceTest.java and the existing 

https://git.onap.org/policy/drools-pdp/tree/policy-management/src/test/java/org/onap/policy/drools/persistence/test/SystemPersistenceTest.java

The SystemPersistenceTest is just the interface and FileSystemPersistenceTest is one (the only one as of now) implementation.   Is it possible to consolidate all tests in one, I think there's a bit of redundancy.

Patch-set: 1
Label: Verified=0