Update patch set 1
Patch Set 1:

One thing that I did discover is the 'settings.xml' file used when running the Junit tests is not necessarily the same one used while compiling. If the '-s' option is used to pass a 'settings.xml' file to 'mvn', the Drools container could be dynamically loading artifacts and dependencies using a different 'settings.xml' file, presumably using the default rules. This may be related to what is happening here.

By the way, I also did another test, where I removed changed the 'drools-artifact1' pom.xml files by removing the 'parent' element, but leaving 'groupId'. That did compile correctly, and may be a workaround for this problem.

Patch-set: 1