Update patch set 1
Patch Set 1:

Ralph, not sure about the possibility of providing a custom settings.xml.   

I think it is worth though to give submit it to gerrit for a quick try: removing the oparent section on the test artifacts but keeping in our root pom the release version change, to see if it passes the "verify", we could abandon that one if does not pass the verify.  

Is it something you may have time to do?

Patch-set: 1
Label: Verified=0