Update patch set 4
Patch Set 4: Code-Review-1

There should be parallel review in docker repo to realize this functionality, but first, I would like to understand what this functionality is trying to achieve to give better comments.   Can you add some additional information in POLICY-341, which perhaps could be added to the wiki page for this feature?   

Obviously, it just monitors an additional nexus repository, but is not clear to me for what purpose?   There is no functionality to switch nexus repositories on the fly on the drools pdp.   So, I would like to understand the consequences of adding another default repository, I suppose this will only apply in a context of pdp-d being used in an active/standby context.

Patch-set: 4
Label: Code-Review=-1