Update patch set 1
Patch Set 1: -Code-Review

> (11 comments)

L144: It installs and runs.  Wouldn't that indicate that it has all of the dependencies that it needs?

L88: Is it ok to subclass from RuntimeException, or will sonar complain about that, too?

L838: Is it sufficient to simply add a comment to an empty block of code?  Or does sonar require the block to actually contain an executable line of code in every block?

Patch-set: 1