Update patch set 1
Patch Set 1:

L144 - yes, it is good
L88 - you can use and existing subclass if it fits the purpose, for example a commonly used one is IllegalArgumentException or create a custom one for your semantics.
L838 - sonar requires to log exceptions, passing explicitly "e" to the logger, or rethrow them.   You could set the logging statement if you don't care to trace or debug, just to satisfy sonar.  alternative is to mark it as false positive but that requires manual maintenance

Patch-set: 1
Label: Verified=0