Update patch set 1
Patch Set 1: Code-Review-1

The issue is that qa default ONAP deployment will not have a secondary nexus repo, but the settings.xml used will have an invalid entry after installation in the settings.xml.

I'm open for discussion but I tend to think that the right way to handle this functionality is to attach an "install" script to the appropriate feature, I think it may be the active-standby feature.

There are some minor changes in the base.conf associated with that but we can discuss offline if there's agreement.

Patch-set: 1