Update patch set 1
Patch Set 1: Code-Review-1

Rather than making two repositories a requirement, it would be best to provide it as an option. Note that this means that there would now be a third 'settings.xml' variant to choose from (there is already a second 'standalone-settings.xml' file), and an alternate 'base.conf' to choose from.

I have been thinking of a generic way to make this kind of selection. Suppose there were a directory called 'settings.xml.select' which had three files in it: '01_twoRepositories', '02_singleRepository', '03_standAlone', which contains the three variants. The '${{VARNAME}}' variables are expanded in each of the files, and the one actually selected would be the first one that doesn't have any unexpanded '${{VARNAME}}' variables.

Note that this could be a generic mechanism -- any directory of the form 'xyzzy.select' could contain multiple candidates for the file 'xyzzy'. This would also allow a more generic 'configure' function in 'docker-install.sh', and eliminate the specialized 'configure_settings' function.

Patch-set: 1
Label: Code-Review=-1