Update patch set 1
Patch Set 1:

OK, there is agreement then to take this back to design phase.   With regards on how it's done need more details on the proposals.   One important decision is if this is a base only feature (all the work in installation), or specifically associated with a feature (work on feature install hooks), or both.

Patch-set: 1