Update patch set 2
Patch Set 2:

> Is this related to the other gerrit review https://gerrit.onap.org/r/#/c/25781/?
 > Should one be done before the other? Or was more work going to
 > happen on the other one?

This is basically the same code as the other review, with some changes to get it to work in this location.  Jorge asked that the changes be moved into drools-pdp. I created a new review for it because I don't know how to do reviews across projects.

Patch-set: 2