Update patch set 1
Patch Set 1:

We were told in the past not to have the JIRA ticket in the first line to have enough characters for the description, but have it in the Issue-ID (which you have).   Can you remove it from first line?  At the same tine, if you can remove the additional signed-off that leaked into the Issue-ID line?   Thanks!

Patch-set: 1
Label: Verified=0