feature interception of facts before inject
Added interceptor to drools controller, post-decoding and
pre/after injection into kie sessions.   The decoded
object is passed as is.   The set of hooks will be
expanded in the future.

This is added in the context of logging for POLICY-558
to track control loop transactions, per logging requirements.
The idea is that VirtualControlLoopNotifications on the
drools-applications area will use it to track control loop
transactions, times, etc .. to satisfy logging requirements.
This approach also reliefs the drools applications logic, (.drl)
for dealing with this.   The logic therefore will be isolated into
a single feature.

Change-Id: Idf840c528e6175be4aaee3bf114091c203b37bc8
Issue-ID: POLICY-558
Signed-off-by: Jorge Hernandez <jh1730@att.com>