Update patch set 2
Patch Set 2: Code-Review-1

What is the benefit of doing this name change?   

There are side effects for those doing testing in our team and integration teams.   OOM and integration source code repos I think have occurrences and may need changes as well. I think there are references to default naming.

Daniel Cruz in our team and integration teams need to be aware, so they are not picking up stale images on their testing.   We need to make sure that the change is socialized with testing teams so they don't test with old images, if they don't do appropriate changes in their lab configurations.   It may be better to do these changes in the beginning of a release.

What activity is publishing to github?   May be somebody that have been experimenting.  I only see v1.1.0 and v1.1.1 so it may have been a short lived experiment with wrong scripts somehow and somewhere.   It does not seem clear to me that these scripts are the source of the problem, nor I understand why those images have been pushed to dockerhub.

Patch-set: 2