Update patch set 2
Patch Set 2:

Snae issue exists with msb images, they get prepended  "msb-".   The script that is publishing to dockerhub probably flattens >2 level directory style naming hierarchies into 2-level ones >2 ie. in onap  "onap/y/z" to "onap/y-z" blindly for dockerhub.

Are there naming restrictions in dockerhub?   In any case, dockerhub, doesn't even have the official amsterdam release which is v1.1.3.   

The change to have unified naming structure makes sense if we officially are indeed required to change to dockerhub, but it would be much better to do this at the very beginning of a release, not towards the end of M4.

Patch-set: 2