Update patch set 1
Patch Set 1:

OK - I'm completely missing the "big picture" on this. How does this relate to Drew's 'LockFeatureAPI' interface, and what is the purpose to either one of these API's? The existing feature API's tend to provide hooks where you can change behavior or just do something when state changes occur on long-lived objects, such as the Drools engine, or a Drools session.

What kind of things would you do with this API? If the intention is to give the ability to provide an alternate lock implementation, it seems that this is not sufficient to do that, because there are additional concepts that you might need, such as an asynchronous notification interface involving callbacks and/or Drools event (blocking the Drools session thread is NOT a good long-term solution).

Patch-set: 1