Update patch set 1
Patch Set 1: Code-Review-1

> OK - I'm completely missing the "big picture" on this. How does
 > this relate to Drew's 'LockFeatureAPI' interface, and what is the
 > purpose to either one of these API's? The existing feature API's
 > tend to provide hooks where you can change behavior or just do
 > something when state changes occur on long-lived objects, such as
 > the Drools engine, or a Drools session.
 > 
 > What kind of things would you do with this API? If the intention is
 > to give the ability to provide an alternate lock implementation, it
 > seems that this is not sufficient to do that, because there are
 > additional concepts that you might need, such as an asynchronous
 > notification interface involving callbacks and/or Drools event
 > (blocking the Drools session thread is NOT a good long-term
 > solution).

This replaces Drew's LockFeatureAPI, as it needs to be in a jar that is always on the classpath, while distributed locking will only be on the classpath if the feature is enabled.  It is intended to work like the PolicyControllerAPI, which has beforeOffer & afterOffer methods to provide hooks before and after events are processed.  In this case, it provides hooks before and after locks are acquired.

For now, PolicyGuard will use it directly.  Although, I could imagine creating a class, here, that provides some of the basic functionality that PolicyGuard currently provides.  ** Should I go ahead and do that?  And if so, should this project be renamed from api-resource-locks to simply resource-locks, since it would now contain more than just an API?

Regardless, to facilitate future use, I can add a callback to the beforeLock and afterLock methods.  (Adding it in both, as the implementer may prefer it in one or the other.)

Patch-set: 1
Label: Code-Review=-1