Update patch set 3
Patch Set 3:

> Not sure what the policy-endpoints module project changes is trying
 > to accomplish but seems unrelated to POLICY-748.   For those
 > changes, let's open a new jira ticket that explains what is trying
 > to accomplish and move it out from this review.

The changes are required to support pooling, both the version in Beijing as well as the version planned for Casablanca.  It is a loophole that was discovered during testing.  If you'd prefer, the changes to policy-endpoints can be made under a separate JIRA bug fix, but that would have to be merged before this JIRA is merged.

Patch-set: 3