Update patch set 1
Patch Set 1:

Jim, I think we need to do something about it.   As it stands right now, I think there may be a leak of unclosed consumers when the setFilter is invoked.  One solution could be to pass the filter in the constructor, another one that does not require passing it in the constructor, is to delay building the consumer until the first fetch.   I think we should open a ticket to track this, I think there's a resource leak.

Patch-set: 1