Update patch set 5
Patch Set 5:

Folks, just to note. If you think something in sonar is false positive or really should not be changed. No problem, just put the comment in the sonar.onap.org and tell either Jorge or mysel. If we agree then we can go and mark them appropriately to clear it. I see a lot of value in the sonar, but agree in that some things are ridiculous.

Patch-set: 5