skip persisted config files when upgrading
add enable and disable scripts to the feature-distributed-locking
to upgrade/downgrade their databases when these features are
enabled or disabled.

allow docker-install.sh and features scripts to override configuration
when the configuration files have been persisted.  These files are
persisted in the context of an OOM installation to save configuration
information across upgrades.

Change-Id: Iff9368c915f6da7f7c75d9a5a20ffe27e290e7d3
Issue-ID: POLICY-795
Signed-off-by: Jorge Hernandez <jh1730@att.com>