Regenerate policy keystore
The keystore has been regenerated using the latest tooling
from AAF.

Previous versions seem to add a bogus " fqdn" which makes
the the particular policy.api.simpledemo.onap.org entry invalid.
The latest version does not append it.

This change is done for consistency with policy/engine for
which proper format of this entry is necessary to interact
with CLAMP.

Drools is never accessed with this identity, therefore the
changes are complimentary and do not require release of new
artifacts.

Change-Id: Ibf6c2e7130dda247ac320f334765e04d369d6b20
Issue-ID: POLICY-1193
Signed-off-by: Jorge Hernandez <jh1730@att.com>