Env. pointing to logging config
Per current logging spec, the location is recommended to be
under /etc/onap/<component>/subcomponent but is acceptable
to have an enviroment variable pointing to it instead.

Environment variable is preferred over fixed location so
the config files are not all over the place, and is not
placed under directories owned by root.

Change-Id: Id5679f49a726fa43147ffb3e141acec30ebfdaaf
Issue-ID: POLICY-1294
Signed-off-by: Jorge Hernandez <jorge.hernandez-herrero@att.com>