Update patch set 5
Patch Set 5:

(1 comment)

There is also this change in policy/commons.  If you can add with another review, to remove the custom jackson parser property name, so to make sure that nobody is misled to using it as supported?

Bigger issue is that the jackson property and other properties should not be there in policy/common since they are not endpoint related.

policy-endpoints/src/main/java/org/onap/policy/common/endpoints/properties/PolicyEndPointProperties.java

Patch-set: 5