Update patch set 1
Patch Set 1:

Just some questions for my understanding, 
is it possible to do it without requiring defining additional properties?   It is kind of a property that would be difficult to understand and configure by folks, so better to hide it to some degree if possible.

What are the possible values?   
What is the default if left unset?   How will it behave?

Patch-set: 1