Update patch set 1
Patch Set 1:

> Just some questions for my understanding,
 > is it possible to do it without requiring defining additional
 > properties?   It is kind of a property that would be difficult to
 > understand and configure by folks, so better to hide it to some
 > degree if possible.
 > 
 > What are the possible values?
 > What is the default if left unset?   How will it behave?

The problem is this: until all code using the http/jersey server classes has been converted to work with gson, things will break if we default to gson right away.  Policy/engine has a lot of these, which may take a while to convert, thus the property is added to each server, as it has been converted to work with gson.  Once EVERYTHING has been converted to work with gson, then that can be made the default and the property can be removed.

That was the plan.  Is there a different approach you'd prefer?  Policy/engine has a lot of servers that have to be converted.  It MAY be possible to do it all in a week, but not sure.  If we went ahead with that, then we wouldn't need these temporary properties, but I don't have time to do any more with this if I'm to make progress on POLICY-1444.

Patch-set: 1