Update patch set 2
Patch Set 2:

I'm just putting this review up to get some opinions. Currently the way this works is based on configuring a logger through a logback.xml extension file for each controller. That means whenever you add a new controller you will have to add a new configuration for it if you want it to work with the feature. It also needs to be named the same as the controller. This may not be as scalable but the benefit is you have complete flexibility on the logger configurations. Another approach would be to programmatically create the loggers so that the system takes care of the naming of the loggers and no configuration needs to be added every time a new controller is added. This comes with a loss of flexibility of configuration. Any preferences?

Patch-set: 2