Update patch set 5
Patch Set 5:

> I tend to prefer #1, although I am not sure if I understand
 > correctly number 2.   An operator should have control to change
 > configuration on the fly, and you may want a mix and match, you may
 > only want to do controller logging for one controller.
 > 
 > One question, right now, if I change the logback-controller.xml,
 > will it reconfigure automatically?   I would assume so, as part of
 > logback functionality, but just checking.

Right, that is the benefit of option 1, it has complete flexibility and functionality of logback. So everything can be updated on the fly. If loggers are created programmatically as in option 2, a restart would be needed to reconfigure the controller loggers.  

For logback-controller.xml, it gets reconfigured automatically. To preserve this and if you agree option 1 is the way to go, then I will remove logback-controller.xml from this pull request and then make a "logback-amsterdam.xml" and "logback-casablanca.xml" in the feature-amsterdam and feature-casablanca modules in drools-application that will be put in the config directory when enabling the controller logging feature.

Patch-set: 5
Label: Code-Review=0