Update patch set 2
Patch Set 2:

> (3 comments)
 > 
 > nice!  Only need one fix: since all PDPs share the same topic, need
 > to discard messages that aren't target to this particular PDP.

Thanks Jim.  

This raises a question, when we start up and we go passive, and then we get a status update message to be assigned to a group, how the PDP would know if it is for it, or somebody else, since the group/subgroup are not known to the PDP.  One idea is that the pdp-update message could carry the "instance" identifier that we report in the pdp-status message.   I think this would be the only scenario where an unique "host identifier" is needed, subsequent pdp-update messages would be group based.

What do you guys think?

Another note, for the pdp-status, I just populate the "instance" in the message that is the hostname.   This is an unique identifier across pdps.   There are some other fields that won't be populated or populated with the same value, so it is redundant.   I am not sure if those are remains from where we were talking about having a mix of kubernetes and application data embedded in the message or something else.   I believe we decided to move away from kubernetes kind of info, as would be difficult to maintain, and may not be the best practice if avoidable.

Let me know your feedback.

Patch-set: 2