Update patch set 2
Patch Set 2:

@Jorge:
Regarding the first item: the "name" field of the PdpUpdate should contain the name of the PDP.

Regarding the second: yes, it looks like "name" and "instance" are the same.  We should probably stick with "name" and remove "instance" to make it consistent with PdpUpdate.  (Or we could change PdpUpdate, if we think "instance" is more descriptive.)

Patch-set: 2