Update patch set 2
Patch Set 2:

If we are using hostname as the value, then I think "name" is more descriptive. Instance to me would indicate that the value would change if we restart the pdp as it would be a new instance of the PDP on the same host, Am I correct?  or am I completely off here?

Patch-set: 2