Update patch set 2
Patch Set 2:

I think we can have several flavors of released base images, some fatter than others on a hierarchical fashion for development cycle and for release ones.   Just changing the application image what it extends we can make the switch.   Anyways, not on this release, but we should do the work, it is just a sane way to structure the system and dictate rules moving forward.

Patch-set: 2