Update patch set 2
Patch Set 2:

So .. Patch 1 is to be able to find the alpine images, that solves the first problem with csit.   Patch 2 in Dockerfile is to solve the problem that the container comes up in a funky state and holds on in there during instantiation until the jenkins job times out, 90 minutes.   Hopefully this will solve it.

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0