Update patch set 4
Patch Set 4:

> We need to change the ci-management jobs - right now the verify
 > builds on a node that does not have docker.

The verify should pass on a non-docker node with patchset 3. It is because I was building the docker image irrespective of the profile and only included pushing the image within the scope of the "docker" profile. I preferred this way since we'd catch issues before doing the merge, however this does not align with pap and pdp.

Patch-set: 4