Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

Thanks Rashmi for this work, I know it is more work but I am interested in preserving some of the previous functionality, more specifically, generating the install-drools-* file, so to be kept as a regular delivery with or without generating the docker image.   I don't see in the verify job output building the install-drools*.zip package and put it under maven as a delivery (when I grepped for it).   This is used by folks to do standalone installs without docker.   It would also be good to build it to catch some issues with the packaging and developer testing without docker.   Let me know your thoughts.

Patch-set: 7