Update patch set 7
Patch Set 7:

"mvn clean install" was building the install-drools-pdp.zip, to produce an install package that is portable and docker agnostic.  The docker image build ideally should use that output. I would like to not lose functionality here as it was very useful.  I also think is a better way to structure the build.

Patch-set: 7