Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > (1 comment)
>  > 
>  > > Patch Set 7:
>  > >
>  > > "mvn clean install" was building the install-drools-pdp.zip, to
>  > produce an install package that is portable and docker agnostic. 
>  > The docker image build ideally should use that output. I would like
>  > to not lose functionality here as it was very useful.  I also think
>  > is a better way to structure the build.
>  > 
>  > All this JIRA is about is moving to using docker plugin exactly
>  > like the api, pap repos. How is what is happening here different
>  > from those jobs?
>  > 
>  > Keep in mind we have to have subsequent changes to support global
>  > jjb gerrit maven jobs.
> 
> Right, and those repos both have a two-step process: 1) one pom builds the tarball  2) another pom builds the docker image using the tarball.  However, neither of these steps are executed as part of the normal "mvn install" from the top level pom.

The merge job executes this 2 step process - this is the verify job. We can see if there is a different verify job to use for ALL the repos sometime later. Since we need to move to global jjb pretty quickly, I prefer to just make the change to match the other repos. Then we can switch all of them later to a different 2 step verify job after things have settled and we've gotten off the radar for global jjb upgrade.

Patch-set: 7